requirements.txt
fastapi
uvicorn
httpx
supabase
pydantic
python-multipart
python-dotenv
google-genai